feat: convert 'StrictlyPositiveDouble' from inline to normal class
This commit converts the 'StrictlyPositiveDouble' type to normal class for making it compatible with Java code and for improving the readability of the public API binaries.
It also reorders declarations of 'StrictlyPositiveDouble' for grouping them by overrides and type-specific declarations.
This last change makes it easier to read the file without using an Integrated Development Editor (IDE).