chore: reorder configuration of some tasks in documentation plugin
This commit makes a clear separation between the configuration needed for publishing on Maven central and the other one for publishing on our website.
For now, the separation is done through comments, but it will be done through different functions after refactoring the 'TaskContainer.configureDokkaHtml(Project)' function in our documentation's Gradle plugin.