🔧 Update configuration of `checkJvm` Gradle task
This task now depends on the `jvmApiCheck` one if the specified project has the `org.jetbrains.kotlinx.binary-compatibility-validator` Gradle plugin.
Due to this change, this commit also removes the `apiCheck` task from the development group (available by running the `devTasks` task). It is now recommended to use the `checkJvm` task instead.